rag_crew/pyproject.toml
[project]
name = "rag_crew"
version = "0.1.0"
description = "Rag_Crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "google-generativeai>=0.3.0,<1.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
rag_crew = "rag_crew.main:run"
run_crew = "rag_crew.main:run"
train = "rag_crew.main:train"
replay = "rag_crew.main:replay"
test = "rag_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
